feat: add some setup for columns
I am unsure where to continue to modify the format of the display of the data. Would we simply want the python dictionary representing the history to be displayed down the coloumns in the csv OR are we to have a new column representing the move history (state) of each player in the game as it continues. If the later is the case, then we will need to realize that this CSV file will very quickly become GB of data once finished. To avoid this strain on my current system, I request clarification of what we are hoping to accomplish.