docs: Updated the documentation
Update the repository README file to reflect more of the current project progress. Also update each games readme to inclulde explicit examples of illegal moves. Finally add the beginning of section 7 in the overview notebook which reviews what Dr. Penland had mentioned to us in our last meeting/further described in an email.

These updates should be beneficial and help keep everyone on the same page as far as collaboration is concerned.