feat: Update the Paths notebook
The notebook works through even and odd path graphs. We create a few functions to help automate this should we want to build a script that would write to a csv. Additionally, the notebook now has a complete table of contents to allow for easy access to each of the sections.